requirements.txt
Django == 1.9.4
python_dateutil == 2.5.3
unicode_slugify == 0.1.3
